+++ src/migrations/m151125_140002_init.php
@@ -3,7 +3,7 @@
 namespace davidhirtz\yii2\skeleton\migrations;
 
 use davidhirtz\yii2\skeleton\auth\rbac\OwnerRule;
-use davidhirtz\yii2\skeleton\db\MigrationTrait;
+use davidhirtz\yii2\skeleton\db\traits\MigrationTrait;
 use davidhirtz\yii2\skeleton\models\AuthClient;
 use davidhirtz\yii2\skeleton\models\Session;
 use davidhirtz\yii2\skeleton\models\UserLogin;