+++ src/models/traits/StatusAttributeTrait.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace davidhirtz\yii2\skeleton\db;
+namespace davidhirtz\yii2\skeleton\models\traits;
 
-use JetBrains\PhpStorm\Deprecated;
 use Yii;
 
 /**