+++ src/modules/admin/widgets/grids/traits/StatusGridViewTrait.php
@@ -3,7 +3,7 @@
 namespace davidhirtz\yii2\skeleton\modules\admin\widgets\grids\traits;
 
 use davidhirtz\yii2\skeleton\db\ActiveRecord;
-use davidhirtz\yii2\skeleton\db\StatusAttributeTrait;
+use davidhirtz\yii2\skeleton\models\traits\StatusAttributeTrait;
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\widgets\bootstrap\ButtonDropdown;
 use davidhirtz\yii2\skeleton\widgets\fontawesome\Icon;
@@ -61,7 +61,7 @@ trait StatusGridViewTrait
     }
 
     /**
-     * @param StatusAttributeTrait $model
+     * @param \davidhirtz\yii2\skeleton\models\traits\StatusAttributeTrait $model
      * @return Icon
      */
     protected function getStatusIcon($model)
@@ -93,7 +93,7 @@ trait StatusGridViewTrait
     }
 
     /**
-     * @param StatusAttributeTrait|null $model
+     * @param \davidhirtz\yii2\skeleton\models\traits\StatusAttributeTrait|null $model
      * @return array
      */
     protected function statusSelectionButtonItems($model = null): array